streamlit/requirements.txt
# Dependencias principales
streamlit==1.47.0
mysql-connector-python==9.4.0
pandas==2.3.1
plotly==6.2.0
redis==9.4.0
python-dotenv==1.1.1

# Visualización y análisis
matplotlib==3.10.3
seaborn==0.13.2
altair==5.3.0

# Utilidades adicionales
numpy==1.26.4
pytz==2024.1
tabulate==0.9.0